requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
playwright>=1.40.0
